:recycle: Merge ConvertionUtils to ParseUtils
