+++ api/src/utils/parseUtils.ts
@@ -12,4 +12,26 @@ export class ParseUtils {
             ? undefined
             : parseFloat(value);
     }
+
+    static parseBoolean(
+        str: string,
+        default_value: any = undefined
+    ): boolean {
+        const truth_values = ["1", "true"];
+        const false_values = ["0", "false"];
+
+        if (ValidationUtils.isEmpty(str)) {
+            return default_value;
+        }
+
+        const lower_str = str.toString().toLowerCase();
+
+        if (truth_values.includes(lower_str)) {
+            return true;
+        } else if (false_values.includes(lower_str)) {
+            return false;
+        } else {
+            return default_value;
+        }
+    }
 }
